# api_contract.yaml
openapi: 3.0.0
info:
  title: PlantSaathi API
  version: 1.0.0
  description: |
    PlantSaathi is a Flutter-based mobile app that helps users identify plants, learn about them, and ask AI-based plant care questions.

    # NOTE FOR FRONTEND DEVELOPERS:
    # All API responses are wrapped as:
    #   Success: { "success": true, "data": ... }
    #   Error:   { "success": false, "message": "..." }
    # Always check the 'success' field before using 'data'.
    # For AI, if answer is "AI service busy, try again later", show fallback UI.
    # Protected routes require: Authorization: Bearer <token>

servers:
  - url: https://api.plantsaathi.com/v1

paths:
  /plants/search:
    get:
      tags: [Plants]
      summary: Search plants by name
      parameters:
        - name: q
          in: query
          description: Search query for plant name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of matching plants
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      plants:
                        type: array
                        items:
                          $ref: '#/components/schemas/PlantSummary'
                      pagination:
                        type: object
                        properties:
                          page:
                            type: integer
                          limit:
                            type: integer
                          total:
                            type: integer
              examples:
                frontend:
                  value:
                    success: true
                    data:
                      plants:
                        - id: "1"
                          common_name: "Rose"
                          scientific_name: "Rosa"
                          image_url: "https://example.com/rose.jpg"
                        - id: "2"
                          common_name: "Tulsi"
                          scientific_name: "Ocimum tenuiflorum"
                          image_url: "https://example.com/tulsi.jpg"
                      pagination:
                        page: 1
                        limit: 10
                        total: 2
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
              example:
                success: false
                message: "Missing required query parameter: q"
        '404':
          description: No plants found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
              example:
                success: false
                message: "No plants found"

  /plants/{id}:
    get:
      tags: [Plants]
      summary: Get detailed plant information
      parameters:
        - name: id
          in: path
          required: true
          description: Plant ID
          schema:
            type: string
      responses:
        '200':
          description: Plant details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Plant'
              examples:
                example:
                  value:
                    success: true
                    data:
                      id: "1"
                      common_name: "Rose"
                      scientific_name: "Rosa"
                      uses: "Ornamental, medicinal"
                      benefits: "Aesthetic, fragrance"
                      where_it_grows: "Gardens, parks"
                      how_to_grow: "Plant in well-drained soil, water regularly"
                      image_url: "https://example.com/rose.jpg"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
              example:
                success: false
                message: "Invalid or missing id parameter"
        '404':
          description: Plant not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
              example:
                success: false
                message: "Plant not found"

  /plants/identify:
    post:
      tags: [Plants]
      summary: Identify plant using an image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: Image file to identify
      responses:
        '200':
          description: Identified plant details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Plant'
              examples:
                example:
                  value:
                    success: true
                    data:
                      id: "2"
                      common_name: "Tulsi"
                      scientific_name: "Ocimum tenuiflorum"
                      uses: "Medicinal, religious"
                      benefits: "Immunity booster, air purifier"
                      where_it_grows: "Tropical regions, home gardens"
                      how_to_grow: "Needs sunlight, water moderately"
                      image_url: "https://example.com/tulsi.jpg"
        '400':
          description: Bad request (e.g., missing image)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Plant not identified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Identification failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /ai/ask:
    post:
      tags: [AI]
      summary: Ask AI questions about plants
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIQuestion'
            examples:
              example:
                value:
                  question: "How often should I water a rose plant?"
      responses:
        '200':
          description: AI-generated answer
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      answer:
                        type: string
              examples:
                example:
                  value:
                    success: true
                    data:
                      answer: "Water rose plants every 2-3 days, keeping the soil moist but not soggy."
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /favorites:
    get:
      tags: [Favorites]
      summary: Get list of saved favorites
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of favorites
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Favorite'
              examples:
                example:
                  value:
                    success: true
                    data:
                      - id: "fav1"
                        type: "plant"
                        plant_id: "1"
                      - id: "fav2"
                        type: "ai"
                        text: "Water rose plants every 2-3 days."
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /favorites/{id}:
    delete:
      tags: [Favorites]
      summary: Remove a saved favorite
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Favorite id
          schema:
            type: string
      responses:
        '200':
          description: Favorite removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Favorite'
              examples:
                example:
                  value:
                    success: true
                    data:
                      id: "fav1"
                      type: "plant"
                      plant_id: "1"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Favorite not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Favorites]
      summary: Save a favorite plant or AI response
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FavoriteInput'
            examples:
              plantExample:
                value:
                  type: "plant"
                  plant_id: "2"
              aiExample:
                value:
                  type: "ai"
                  text: "Tulsi helps boost immunity."
      responses:
        '201':
          description: Favorite saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Favorite'
              examples:
                plantExample:
                  value:
                    success: true
                    data:
                      id: "fav3"
                      type: "plant"
                      plant_id: "2"
                aiExample:
                  value:
                    success: true
                    data:
                      id: "fav4"
                      type: "ai"
                      text: "Tulsi helps boost immunity."
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /recommendations:
    get:
      tags: [Recommendations]
      summary: Get personalized plant recommendations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Recommendation list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PlantSummary'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PlantSummary:
      type: object
      required:
        - id
        - common_name
        - scientific_name
      properties:
        id:
          type: string
        common_name:
          type: string
        scientific_name:
          type: string
        image_url:
          type: string

    Plant:
      allOf:
        - $ref: '#/components/schemas/PlantSummary'
        - type: object
          required:
            - uses
            - benefits
            - where_it_grows
            - how_to_grow
          properties:
            uses:
              type: string
            benefits:
              type: string
            where_it_grows:
              type: string
            how_to_grow:
              type: string

    AIQuestion:
      type: object
      properties:
        question:
          type: string
      required:
        - question

    AIResponse:
      type: object
      required:
        - question
        - answer
      properties:
        question:
          type: string
        answer:
          type: string

    Favorite:
      type: object
      description: |
        Favorite item. If type = "plant", plant_id must be present. If type = "ai", text must be present.
      required:
        - id
        - type
      properties:
        id:
          type: string
        type:
          type: string
          enum: [plant, ai]
        plant_id:
          type: string
          nullable: true
        text:
          type: string
          nullable: true

    FavoriteInput:
      type: object
      properties:
        type:
          type: string
          enum: [plant, ai]
        plant_id:
          type: string
          nullable: true
        text:
          type: string
          nullable: true
      required:
        - type

    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message