# backend/src/routes/favorite.routes.js
const express = require('express');
const router = express.Router();
const favoriteController = require('../controllers/favorite.controller');
const auth = require('../middlewares/auth.middleware');


/**
 * @swagger
 * /favorites:
 *   get:
 *     summary: Get user's favorites
 *     parameters:
 *       - in: query
 *         name: page
 *         schema:
 *           type: integer
 *         required: false
 *         description: Page number
 *       - in: query
 *         name: limit
 *         schema:
 *           type: integer
 *         required: false
 *         description: Results per page
 *     responses:
 *       200:
 *         description: List of favorites
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 data:
 *                   type: array
 *                   items:
 *                     $ref: '#/components/schemas/Favorite'
 *                 pagination:
 *                   type: object
 *                   properties:
 *                     page:
 *                       type: integer
 *                     limit:
 *                       type: integer
 *                     total:
 *                       type: integer
 *   post:
 *     summary: Add a favorite (plant or AI)
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             properties:
 *               type:
 *                 type: string
 *                 enum: [plant, ai]
 *               plant_id:
 *                 type: string
 *               text:
 *                 type: string
 *     responses:
 *       201:
 *         description: Favorite added
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/Favorite'
 *   delete:
 *     summary: Remove a favorite by id
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *         description: Favorite id
 *     responses:
 *       200:
 *         description: Favorite removed
 */
router.get('/', auth, favoriteController.getFavorites);
router.post('/', auth, favoriteController.addFavorite);
router.delete('/:id', auth, favoriteController.deleteFavorite);

module.exports = router;